+++ scripts/autoupdate-bash.sh
@@ -1,4 +1,11 @@
 #!/bin/bash
+
+get_latest_release() {
+	curl --silent "https://api.github.com/repos/$1/releases/latest" |	# Get latest release from GitHub api
+	grep '"tag_name":' |												# Get tag line
+	sed -E 's/.*"([^"]+)".*/\1/'										# Pluck JSON value
+}
+
 set -e $1
 
 opkg update || true
@@ -13,24 +20,18 @@ proceed_command losetup
 proceed_command resize2fs
 opkg install coreutils-truncate || true
 
-board_id=$(cat /etc/board.json | jsonfilter -e '@["model"].id' | sed 's/friendly.*,nanopi-//;s/xunlong,orangepi-//;s/^r1s-h5$/r1s/;s/^r1$/r1s-h3/;s/^r1-plus$/r1p/;s/default-string-default-string/x86/')
+board_id=$(cat /etc/board.json | jsonfilter -e '@["model"].id' | sed 's/friendly.*,nanopi-//;s/xunlong,orangepi-//;s/^r2$/r2s/;s/^r1s-h5$/r1s/;s/^r1$/r1s-h3/;s/^r1-plus$/r1p/;s/^r1-plus-lts$/r1p-lts/;s/default-string-default-string/x86/;s/vmware-inc-vmware7-1/x86/;s/qemu-standard-pc-q35-ich9-2009/x86/;s/qemu-standard-pc-i440fx-piix-1996/x86/')
+arch=`uname -m`
+[ $arch == 'x86_64' ] && board_id='x86';
 mount -t tmpfs -o remount,size=850m tmpfs /tmp
 rm -rf /tmp/upg && mkdir /tmp/upg && cd /tmp/upg
 
-md5sum=`wget https://ghproxy.com/https://github.com/klever1988/nanopi-openwrt/releases/download/$(date +%Y-%m-%d)/$board_id$ver.img.gz -O- | tee >(gzip -dc>$board_id.img) | md5sum | awk '{print $1}'`
+latest_release_tag=`get_latest_release stupidloud/nanopi-openwrt`
+echo -e '\e[92m准备更新到'$latest_release_tag'\e[0m'
+md5sum=`wget https://ghproxy.com/https://github.com/stupidloud/nanopi-openwrt/releases/download/$latest_release_tag/$board_id$ver.img.gz -O- | tee >(gzip -dc>$board_id.img) | md5sum | awk '{print $1}'`
 if [ "$md5sum" != "d41d8cd98f00b204e9800998ecf8427e" ]; then
-	wget https://ghproxy.com/https://github.com/klever1988/nanopi-openwrt/releases/download/$(date +%Y-%m-%d)/$board_id$ver.img.gz.md5 -O md5sum.txt
-	echo -e '\e[92m今天固件已下载，准备解压\e[0m'
-else
-	echo -e '\e[91m今天的固件还没更新，尝试下载昨天的固件\e[0m'
-	md5sum=`wget https://ghproxy.com/https://github.com/klever1988/nanopi-openwrt/releases/download/$(date -d "@$(( $(busybox date +%s) - 86400))" +%Y-%m-%d)/$board_id$ver.img.gz -O- | tee >(gzip -dc>$board_id.img) | md5sum | awk '{print $1}'`
-	if [ "$md5sum" != "d41d8cd98f00b204e9800998ecf8427e" ]; then
-		wget https://ghproxy.com/https://github.com/klever1988/nanopi-openwrt/releases/download/$(date -d "@$(( $(busybox date +%s) - 86400))" +%Y-%m-%d)/$board_id$ver.img.gz.md5 -O md5sum.txt
-		echo -e '\e[92m昨天的固件已下载，准备解压\e[0m'
-	else
-		echo -e '\e[91m没找到最新的固件，脚本退出\e[0m'
-		exit 1
-	fi
+	wget https://ghproxy.com/https://github.com/stupidloud/nanopi-openwrt/releases/download/$latest_release_tag/$board_id$ver.img.gz.md5 -O md5sum.txt
+	echo -e '\e[92m'$latest_release_tag'固件已下载\e[0m'
 fi
 
 md5r=`awk '{print $1}' md5sum.txt`
@@ -42,7 +43,7 @@ fi
 mv $board_id.img FriendlyWrt.img
 block_device='mmcblk0'
 [ ! -d /sys/block/$block_device ] && block_device='mmcblk1'
-[ $board_id = 'x86' ] && block_device='sda'
+[ $board_id = 'x86' ] && block_device=${disk:-sda}
 bs=`expr $(cat /sys/block/$block_device/size) \* 512`
 truncate -s $bs FriendlyWrt.img || ../truncate -s $bs FriendlyWrt.img
 echo ", +" | sfdisk -N 2 FriendlyWrt.img