+++ scripts/merge_packages.sh
@@ -1,12 +1,17 @@
 function merge_package(){
     pn=`echo $1 | rev | cut -d'/' -f 1 | rev`
-    find package/ -follow -name $pn -not -path "package/custom/*" | xargs -rt rm -r
+    find package/ -follow -name $pn -not -path "package/custom/*" | xargs -rt rm -rf
     if [ ! -z "$2" ]; then
-        find package/ -follow -name $2 -not -path "package/custom/*" | xargs -rt rm -r
+        find package/ -follow -name $2 -not -path "package/custom/*" | xargs -rt rm -rf
     fi
 
     if [[ $1 == *'/trunk/'* || $1 == *'/branches/'* ]]; then
         svn export $1
+    elif [[ $1 == *'/tree/'* ]]; then
+        repo=`echo $1|awk -F'/tree/' '{print $1}'`
+        git clone --depth=1 --single-branch $repo .tmprepo
+        mv .tmprepo/$pn $pn
+        rm .tmprepo -rf
     else
         git clone --depth=1 --single-branch $3 $1
         rm -rf $pn/.git
@@ -14,7 +19,7 @@ function merge_package(){
     mv $pn package/custom/
 }
 function drop_package(){
-    find package/ -follow -name $1 -not -path "package/custom/*" | xargs -rt rm -r
+    find package/ -follow -name $1 -not -path "package/custom/*" | xargs -rt rm -rf
 }
 function merge_feed(){
     if [ ! -d "feed/$1" ]; then
@@ -25,26 +30,30 @@ function merge_feed(){
     ./scripts/feeds install -a -p $1
 }
 
+rm -rf feeds/packages/net/mosdns package/feeds/packages/mosdns
 rm -rf package/custom; mkdir package/custom
 merge_feed nas "https://github.com/linkease/nas-packages;master"
 merge_feed nas_luci "https://github.com/linkease/nas-packages-luci;main"
-merge_package https://github.com/klever1988/helloworld/branches/tmp/luci-app-ssr-plus
-merge_package https://github.com/klever1988/helloworld/branches/tmp/mosdns
-#merge_package https://github.com/klever1988/openwrt-mos/trunk/luci-app-mosdns
-merge_package https://github.com/project-lede/luci-app-godproxy
-merge_package https://github.com/Beginner-Go/luci-app-tencentddns
-merge_package https://github.com/sundaqiang/openwrt-packages/trunk/luci-app-services-wolplus
-merge_package https://github.com/coolsnowwolf/lede/trunk/package/lean/ntfs3-mount
-merge_package https://github.com/coolsnowwolf/lede/trunk/package/lean/ntfs3
-merge_package https://github.com/coolsnowwolf/lede/trunk/package/lean/uugamebooster
-merge_package https://github.com/coolsnowwolf/lede/trunk/package/lean/luci-app-uugamebooster
-merge_package https://github.com/kuoruan/openwrt-frp frp
-merge_package https://github.com/kuoruan/luci-app-frpc
-merge_package https://github.com/small-5/luci-app-adblock-plus
-merge_package https://github.com/zxlhhyccc/luci-app-v2raya
+rm -r package/feeds/nas_luci/luci-app-ddnsto
+merge_feed helloworld "https://github.com/stupidloud/helloworld;tmp"
+merge_package https://github.com/ilxp/luci-app-ikoolproxy
+merge_package https://github.com/sundaqiang/openwrt-packages/trunk/luci-app-wolplus
 merge_package https://github.com/messense/aliyundrive-webdav/trunk/openwrt/aliyundrive-webdav
 merge_package https://github.com/messense/aliyundrive-webdav/trunk/openwrt/luci-app-aliyundrive-webdav
+merge_package "-b 18.06 https://github.com/jerrykuku/luci-theme-argon"
+merge_package https://github.com/vernesong/OpenClash/trunk/luci-app-openclash
+merge_package https://github.com/NateLol/luci-app-oled
+merge_package https://github.com/xiaorouji/openwrt-passwall-packages/tree/main/brook
+merge_package https://github.com/xiaorouji/openwrt-passwall-packages/tree/main/chinadns-ng
+merge_package https://github.com/xiaorouji/openwrt-passwall-packages/tree/main/trojan-go
+merge_package https://github.com/xiaorouji/openwrt-passwall-packages/tree/main/trojan-plus
+merge_package https://github.com/xiaorouji/openwrt-passwall-packages/tree/main/sing-box
+merge_package "-b main https://github.com/xiaorouji/openwrt-passwall"
+merge_package https://github.com/jerrykuku/lua-maxminddb
+merge_package https://github.com/jerrykuku/luci-app-vssr
+merge_package https://github.com/kongfl888/luci-app-adguardhome
 drop_package luci-app-cd8021x
 drop_package luci-app-cifs
 drop_package verysync
 drop_package luci-app-verysync
+drop_package luci-app-mosdns